refactor(data): STTPermissionService를 STTService에 통합하고 DefaultSTTRepository로 병합
- STTPermissionService 프로토콜 제거 및 checkPermission/requestPermission을 STTService에 통합
- SpeechService의 STTPermissionService 준수 제거 (STTService만 준수)
- DefaultSTTPermissionRepository를 DefaultSTTRepository로 병합 (permissionService 의존성 제거)
- MockSTTPermissionService 제거 및 권한 메서드를 MockSTTService에 통합
- DefaultSTTRepositoryTest에 권한 테스트 통합
- DependencyProvider의 sttPermissionService 프로퍼티 제거
