+++ Data/Sources/Repositories/VoiceNotes/DefaultSTTRepository.swift
@@ -3,7 +3,7 @@ import Domain
 import Foundation
 
 /// 음성 인식(STT) 리포지토리 기본 구현체.
-public struct DefaultSTTRepository: STTRepository {
+public struct DefaultSTTRepository: STTRepository, STTPermissionRepository {
     private let service: any STTService
 
     public init(service: any STTService) {
@@ -19,5 +19,15 @@ public struct DefaultSTTRepository: STTRepository {
             AppLogger.error(error)
             throw STTRepositoryError(error)
         }
+    }
+
+    public func checkSTTPermission() async throws(STTPermissionRepositoryError) -> PermissionStatus {
+        if Task.isCancelled { throw .cancelled }
+        return await service.checkPermission()
+    }
+
+    public func requestSTTPermission() async throws(STTPermissionRepositoryError) -> PermissionStatus {
+        if Task.isCancelled { throw .cancelled }
+        return await service.requestPermission()
     }
 }
